update: atoi

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: jmanani <jmanani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 15:50:22 by jmanani           #+#    #+#             */
-/*   Updated: 2025/10/23 20:32:59 by jmanani          ###   ########.fr       */
+/*   Updated: 2025/10/24 13:49:22 by jmanani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int	calc_num(const char *str, int index)
 
 int	ft_atoi(const char *str)
 {
-	size_t	num;
+	int		num;
 	int		is_neg;
 	size_t	index;
